# openapi.yaml
openapi: 3.0.3
info:
  title: Treasure Shop API
  version: 1.0.0
  description: API documentation for the Treasure Shop backend
servers:
  - url: http://localhost:8080
paths:
...
  /api/orders:
    post:
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order created
        '400':
          description: Bad request
    get:
      summary: List orders (filterable by user_id, writer_id, status)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
          description: Filter by user ID (optional)
        - in: query
          name: writer_id
          schema:
            type: string
          description: Filter by writer ID (optional)
        - in: query
          name: status
          schema:
            type: string
          description: Filter by order status (optional)
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: Page number (default 1)
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Page size (default 10, max 100)
      responses:
        '200':
          description: List of orders (paginated and filterable)
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        user_id:
          type: string
        order_type_id:
          type: string
        order_level_id:
          type: string
          description: Foreign key to OrderLevel (required)
        level_name:
          type: string
          description: Name of the order level (response only)
        order_pages_id:
          type: string
          description: Foreign key to OrderPages (required)
        order_pages_name:
          type: string
          description: Name of the order pages (response only)
        order_urgency_id:
          type: string
          description: Foreign key to OrderUrgency (required)
        order_urgency_name:
          type: string
          description: Name of the order urgency (response only)
        order_style_id:
          type: string
          description: Foreign key to OrderStyle (required)
        order_style_name:
          type: string
          description: Name of the order style (response only)
        title:
          type: string
        description:
          type: string
        price:
          type: number
        status:
          type: string
        writer_id:
          type: string
        submission_date:
          type: string
          format: date-time
        feedback:
          type: string
        is_high_priority:
          type: boolean
          description: Whether the order is high priority (required)
...